feat : Lotto 작성 및 예외처리
