+++ src/main/java/lotto/domain/Lotto.java
@@ -1,0 +1,49 @@
+package lotto.domain;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+public class Lotto {
+    private final List<Integer> numbers;
+
+    public Lotto(List<Integer> numbers) {
+        validateLottoLength(numbers);
+        validateLottoRange(numbers);
+        validateLottoDuplicate(numbers);
+        this.numbers = numbers;
+    }
+
+    public List<Integer> getNumbers() {
+        return numbers;
+    }
+
+
+
+
+
+
+    private void validateLottoLength(List<Integer> numbers) {
+        if (numbers.size() != 6) {
+            throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
+        }
+    }
+
+    private void validateLottoRange(List<Integer> numbers) {
+        for (Integer number : numbers) {
+            if (number < 1 || number > 45) {
+                throw new IllegalArgumentException("로또 번호는 1부터 45 사이의 숫자여야 합니다.");
+            }
+        }
+    }
+
+    //fixme 여기서?
+    private void validateLottoDuplicate(List<Integer> numbers) {
+        Set<Integer> numbersSet = new HashSet<>(numbers);
+        if (numbersSet.size() != numbers.size()) {
+            throw new IllegalArgumentException("중복된 로또 번호가 존재합니다.");
+        }
+    }
+
+    // TODO: 추가 기능 구현
+}
